requirements.txt
pyyaml
reportlab
svglib
beautifulsoup4
fuzzywuzzy
python-Levenshtein